@@ -20,22 +20,24 @@ accelerate = "^0.26.1"
 bitsandbytes = "^0.43.0"
 torch = "2.2"
 peft = "^0.11.0"
-opencv-python = "^4.9.0.80"
+opencv-python = "^4.10"
 einops = "^0.7.0"
 decord = "^0.6.0"
 sentencepiece = "^0.1.99"
 spacy = "3.7.2"
 en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz"}
 en_core_web_md = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1.tar.gz"}
+en_core_web_lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1.tar.gz"}
 scikit-learn = "^1.4.0"
 dataclasses-json = "^0.6.3"
-trl = "0.8.5"
-datasets = "2.19.0"
+trl = "0.9.4"
+datasets = "2.20.0"
 protobuf = "^4.21"
 wandb = "^0.17"
 pympler = "^1.0.1"
 memory-profiler = "^0.61.0"
 pynvml = "^11.5.0"
+optuna = "^3.6.1"
 
 [build-system]
 requires = ["poetry-core"]
