@@ -30,9 +30,12 @@ parser.add_argument("--vlm_name", type=str, default="llava-hf/llava-1.5-7b-hf", 
 parser.add_argument("--eval_partitions", nargs='+', type=str, default=["val", "test"])
 parser.add_argument("--mistake_detection_strategy", type=str, default="heuristic", choices=list(MISTAKE_DETECTION_STRATEGIES.keys()))
 parser.add_argument("--visual_filter_mode", type=str, required=False, choices=[t.value for t in VisualFilterTypes], help="Visual attention filter mode.")
+parser.add_argument("--visual_filter_strength", type=float, required=False, default=1.0, help="Float strength for masks used in visual filters.")
 parser.add_argument("--batch_size", type=int, default=10, help="Batch size for VQA inference.")
 parser.add_argument("--resume_dir", type=str, help="Path to results directory for previous incomplete run of generating frameVQA examples.")
 parser.add_argument("--debug", action="store_true", help="Pass this argument to run on only a small amount of data for debugging purposes.")
+parser.add_argument("--debug_n_examples", type=int, default=250, help="Configure the number of examples per class to generate for debugging purposes.")
+parser.add_argument("--cache_vqa_frames", action="store_true", help="Pass this argument to cache frames in VQA outputs (e.g., to inspect visual filter resuilts). This consumes a lot of disk space for large datasets.")
 args = parser.parse_args()
 
 # Load VLM(s), processors, visual filters, etc. - if multiple GPUs available, use them
@@ -66,8 +69,8 @@ for worker_index in range(n_workers):
 
     if args.visual_filter_mode is not None:
         if VisualFilterTypes(args.visual_filter_mode) == VisualFilterTypes.Contrastive_Region:
-            visual_filter = ContrastiveRegionFilter(device=f"cuda:{worker_index}")
-            nlp = spacy.load('en_core_web_sm')
+            visual_filter = ContrastiveRegionFilter(mask_strength=args.visual_filter_strength, device=f"cuda:{worker_index}")
+            nlp = spacy.load('en_core_web_lg')
         else:
             raise NotImplementedError(f"Visual filter type {args.visual_filter_mode} is not compatible with SuccessVQA!")
         
@@ -85,8 +88,11 @@ if args.resume_dir is None:
     timestamp = datetime.datetime.now()
     this_results_dir = os.path.join(args.task, f"SuccessVQA_{args.task}")
     if args.debug:
-        this_results_dir += f"_debug"
-    this_results_dir += f"_{args.vlm_name.split('/')[-1]}_{timestamp.strftime('%Y%m%d%H%M%S')}"
+        this_results_dir += f"_debug{args.debug_n_examples}"        
+    this_results_dir += f"_{args.vlm_name.split('/')[-1]}"
+    if args.visual_filter_mode is not None:
+        this_results_dir += f"_{args.visual_filter_mode}{args.visual_filter_strength}"
+    this_results_dir += f"_{timestamp.strftime('%Y%m%d%H%M%S')}"
     this_results_dir = os.path.join(RESULTS_DIR, "vqa_mistake_detection", this_results_dir)
     os.makedirs(this_results_dir)
 else:
@@ -114,9 +120,12 @@ for eval_partition in args.eval_partitions:
 
     # Load mistake detection dataset
     if MistakeDetectionTasks(args.task) == MistakeDetectionTasks.CaptainCook4D:
-        eval_datasets = [CaptainCook4DDataset(data_split=eval_partition, debug_n_examples_per_class=20 if args.debug else None) for _ in range(n_workers)]
+        eval_datasets = [CaptainCook4DDataset(data_split=eval_partition, debug_n_examples_per_class=args.debug_n_examples if args.debug else None) for _ in range(n_workers)]
     elif MistakeDetectionTasks(args.task) == MistakeDetectionTasks.Ego4D:
-        eval_datasets = [Ego4DMistakeDetectionDataset(data_split=eval_partition, debug_n_examples_per_class=100 if args.debug else None) for _ in range(n_workers)]
+        eval_datasets = [Ego4DMistakeDetectionDataset(data_split=eval_partition, 
+                                                      mismatch_augmentation=True,
+                                                      multi_frame=True,
+                                                      debug_n_examples_per_class=args.debug_n_examples if args.debug else None) for _ in range(n_workers)]
     else:
         raise NotImplementedError(f"Haven't implemented usage of {args.task} dataset yet!")                                        
         
@@ -135,7 +144,8 @@ for eval_partition in args.eval_partitions:
                                            [this_results_dir] * n_workers,
                                            [n_workers] * n_workers,
                                            list(range(n_workers)),
-                                           [args.batch_size] * n_workers)
+                                           [args.batch_size] * n_workers,
+                                           [args.cache_vqa_frames] * n_workers)
                              )
         # Compile processed data from partitions
         vqa_outputs = []
@@ -154,7 +164,8 @@ for eval_partition in args.eval_partitions:
                                                     cache_dir=this_results_dir,
                                                     n_workers=1,
                                                     worker_index=0,
-                                                    vqa_batch_size=args.batch_size)
+                                                    vqa_batch_size=args.batch_size,
+                                                    cache_frames=args.cache_vqa_frames)
     
     print("Evaluating and saving results...")
 
