@@ -3,21 +3,20 @@ from travel import init_travel
 init_travel()
 
 import argparse
-import concurrent.futures
 import datetime
 import json
 from memory_profiler import profile
 import os
 from pympler.tracker import SummaryTracker
 import shutil
-import torch
 from tqdm import tqdm
 
-from travel.constants import IMAGES_CHUNK_SIZE, CACHE_FREQUENCY
+from travel.constants import IMAGES_CHUNK_SIZE
 from travel.data.utils import split_list_into_partitions
+from travel.data.utils.image import resize_with_aspect, CACHED_FRAME_DIMENSION
 from travel.data.vqa import VQAOutputs
 from travel.data.vqg_learning import load_frameVQA_examples, save_vqg_training_examples, FrameVQAMistakeDetectionExample, VQGTrainingExample
-from travel.model.grounding import VisualFilterTypes, MASK_STRENGTH
+from travel.model.grounding import VisualFilterTypes
 from travel.model.vqa import save_vqa_outputs
 from travel.model.vqg_learning import FrameVQAMistakeDetectionScorer
 
@@ -26,9 +25,11 @@ parser.add_argument("--vqg_directory", type=str, required=True, help="Directory 
 parser.add_argument("--vlm_name", type=str, default="llava-hf/llava-1.5-7b-hf", help="Name or path to Hugging Face model for VLM.")
 parser.add_argument("--partition", type=str, choices=["train", "val", "test"], help="List of partitions to generate data for.")
 parser.add_argument("--visual_filter_mode", type=str, required=False, choices=[t.value for t in VisualFilterTypes], help="Visual attention filter mode.")
+parser.add_argument("--visual_filter_strength", type=float, required=False, default=1.0, help="Float strength for masks used in visual filters.")
 parser.add_argument("--batch_size", type=int, default=52, help="Batch size for VQA inference. Visual filter batch size is configured in `config.yml`.")
 parser.add_argument("--resume_dir", type=str, help="Path to results directory for previous incomplete run of generating frameVQA examples. Can also be used to add another partition of data to existing reuslts directory.")
 parser.add_argument("--track_memory", action="store_true", help="Pass this argument to use `pympler` to print out summaries of memory usage periodically during execution.")
+parser.add_argument("--cache_vqa_frames", action="store_true", help="Pass this argument to cache frames in VQA outputs (e.g., to inspect visual filter resuilts). This consumes a lot of disk space for large datasets.")
 args = parser.parse_args()
 
 if "_debug" in args.vqg_directory:
@@ -44,7 +45,7 @@ if args.resume_dir is None:
     timestamp = datetime.datetime.now()
     this_results_dir = os.path.join(args.vqg_directory, f"VQA_data_{args.vlm_name.split('/')[-1]}")
     if args.visual_filter_mode is not None:
-        this_results_dir += f"_{args.visual_filter_mode}{MASK_STRENGTH}"
+        this_results_dir += f"_{args.visual_filter_mode}{args.visual_filter_strength}"
     this_results_dir += f"_{timestamp.strftime('%Y%m%d%H%M%S')}"
 else:
     this_results_dir = args.resume_dir
@@ -69,6 +70,7 @@ cache_dirs = [cd for cd in cache_dirs if not os.path.exists(os.path.join(cd, f"v
 # Get ready to run VQA scoring
 def run_vqa_scoring_on_chunk(scorer: FrameVQAMistakeDetectionScorer,
                              frameVQA_examples_chunk: list[FrameVQAMistakeDetectionExample],
+                             cache_dir: str,
                              cache_path: str) -> tuple[list[VQGTrainingExample], list[VQAOutputs]]:
     """Local method to run VQA scoring on a chunk of data."""
     
@@ -88,8 +90,13 @@ def run_vqa_scoring_on_chunk(scorer: FrameVQAMistakeDetectionScorer,
     for outputs in this_vqa_outputs:
         for output in outputs:
             if type(output.frame) != str:
+                output.frame = resize_with_aspect(output.frame, CACHED_FRAME_DIMENSION) # Since this is just for inspection purposes, save a smaller copy
                 frames_to_close.append(output.frame)
-                output.cache_frame(this_results_dir)
+                if args.cache_vqa_frames:
+                    output.cache_frame(cache_dir)
+                else:
+                    # Setting frame to a string prevents frame from being saved in memory later
+                    output.frame = ""
         
     # Close frames - have to do this afterward since some are shared between VQAOutputs
     for frame in frames_to_close:
@@ -116,6 +123,7 @@ def run_vqa_scoring_on_chunks(scorer: FrameVQAMistakeDetectionScorer,
             os.makedirs(cache_dir)
         this_vqg_training_examples, this_vqa_outputs = run_vqa_scoring_on_chunk(scorer=scorer,
                                                                                 frameVQA_examples_chunk=frameVQA_examples_chunk,
+                                                                                cache_dir=cache_dir,
                                                                                 cache_path=os.path.join(cache_dir, "VQA_cache.pt"))
         if args.track_memory:
             print("\nMemory (after chunk)")
@@ -145,6 +153,7 @@ else:
 
     scorer = FrameVQAMistakeDetectionScorer(args.vlm_name,
                                             visual_filter_type=VisualFilterTypes(args.visual_filter_mode) if args.visual_filter_mode is not None else None,
+                                            visual_filter_strength=args.visual_filter_strength,
                                             vlm_device=0,
                                             visual_filter_device=0 if args.visual_filter_mode is not None else None)
     
